refactor(catalog): extract metadata schema persistence
